Factor out getPurchaseURLCallback
